Build system and other updates
* Imagine: Enable ARM64 Android build by default
* Imagine: Don't detect ANDROID_NDK_PATH via CC, instead require ANDROID_NDK_PATH and derive toolchain paths from it
* Imagine: Build fixes when compiling with Android NDK's clang
* Imagine: Update docs
* C64 & MSX: Fix build errors due to FilePicker changes
* Add android-15.mk to emulators for building ARM64 & X86 APK
* Clean up various makefiles

+++ imagine/include/imagine/util/strings.h
@@ -28,7 +28,7 @@ BEGIN_C_DECLS
 
 int hexToInt(char c);
 int numCharInString(const char *s, int search);
-uchar string_hasDotExtension(const char *s, const char *extension);
+int string_hasDotExtension(const char *s, const char *extension);
 int charIsDrawableAscii(int c);
 int charIsDrawableUnicode(int c);
 void string_copyUpToLastCharInstance(char *dest, const char *src, char c);
